Move archive creation into cset bake
This means the link will work for command-line invocations, and it helps
simplify the run_cset_recipe.py to be a simple wrapper script.
