+++ tests/test_execute_recipe.py
@@ -15,6 +15,7 @@
 """Tests for running CSET operator recipes."""
 
 import json
+import zipfile
 from pathlib import Path
 
 import pytest
@@ -102,3 +103,18 @@ def test_execute_recipe_skip_write_metadata_written(tmp_path: Path):
     with open(tmp_path / "meta.json", "rb") as fp:
         metadata = json.load(fp)
     assert metadata["skip_write"]
+
+
+def test_create_diagnostic_archive(tmp_path):
+    """Create ZIP archive of output."""
+    # Create dummy output files.
+    files = {"one", "two", "three"}
+    for filename in files:
+        (tmp_path / filename).touch()
+
+    CSET.operators.create_diagnostic_archive(tmp_path)
+    archive_path = tmp_path / "diagnostic.zip"
+    assert archive_path.is_file()
+    with zipfile.ZipFile(archive_path, "r") as archive:
+        # Check all files are now in archive.
+        assert set(archive.namelist()) == files
